Dirty commit, sorry.
With this commit we finally have a working network game! 
Paddles can be moved, and the players will see the updated position. 

Now we should implement the ball logic and we also need to limit the 
server's loops/refresh rate.

There are some things that need to be cleaned: there is a lot of copy-pasted code, mainly in the RemotePlayerInput class and in MultiplayerPong (lines 53-70).